+++ R/das_effort.R
@@ -11,7 +11,8 @@
 #'   These values must be column names from the output of \code{\link{das_process}},
 #'   e.g. 'Bft', 'SwellHght', etc.
 #'   If \code{method == "condition"}, then these also are the conditions which
-#'   trigger segment chopping when they change.
+#'   trigger segment chopping when they change. Only the following conditions can be used for chopping:
+#'   'Bft', 'SwellHght', 'RainFog', 'HorizSun', 'VertSun', 'Glare', 'Vis', 'Course', 'SpdKt'
 #' @param strata.files list of path(s) of the CSV file(s) with points defining each stratum.
 #'   The CSV files must contain headers and be a closed polygon.
 #'   The list should be named; see the Details section.